[pipeline] Add basic pipeline class.
